Refactored helpers into subfolders
