+++ orion/ui/helpers/menu.js
@@ -2,6 +2,7 @@ import React, {PropTypes} from "react";
 import classNames from "classnames/bind";
 import {Link} from 'react-router';
 import {slugify} from 'orion/utils/string';
+import {Div} from './div';
 
 export const VMenuLink = ({to, index, hash, children})=> {
     if (to) {
